Move all .c and .h files into Go package directories
