+++ fiftyonepattern/cache.c
