+++ fiftyonepattern/city.c
