+++ fiftyonetrie/51Degrees.i
@@ -65,7 +65,7 @@
 /*
 #cgo LDFLAGS: -lpthread
 #include <time.h>
-#include "../threading.c"
+#include "threading.c"
 #include "../cache.c"
 */
 import "C"
