+++ fiftyonetrie/FiftyOneDegreesTrieV3.go
@@ -16,7 +16,7 @@ package fiftyonetrie
 /*
 #cgo LDFLAGS: -lpthread
 #include <time.h>
-#include "../threading.c"
+#include "threading.c"
 #include "../cache.c"
 */
 import "C"
