--- go.mod
@@ -1,3 +1,0 @@
-module github.com/Kochava/fiftyonego
-
-go 1.13
